Fix differences duplication in dict compare
`.VisitingIndex` makes a copy of the entire context, which then gets merged with the same original context, duplicating the `Differences`. In case a difference was detected at the start of a dictionary, you end up with `2^keys` amount of duplicate differences.

Fix this by creating a new empty context for the inner comparison.